+++ crates/sparse-ngrams/src/lib.rs
@@ -43,120 +43,3 @@ pub const MAX_SPARSE_GRAM_SIZE: u32 = 8;
 
 pub use extract::{collect_sparse_grams, collect_sparse_grams_deque, collect_sparse_grams_scan, max_sparse_grams};
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn collect_to_vec(content: &[u8], f: fn(&[u8], &mut [NGram]) -> usize) -> Vec<NGram> {
-        let mut buf = vec![NGram::from_rolling_hash(0, 0); max_sparse_grams(content.len())];
-        let count = f(content, &mut buf);
-        buf.truncate(count);
-        buf
-    }
-
-    #[test]
-    fn test_empty_input() {
-        assert!(collect_sparse_grams(b"").is_empty());
-    }
-
-    #[test]
-    fn test_single_byte() {
-        assert!(collect_sparse_grams(b"a").is_empty());
-    }
-
-    #[test]
-    fn test_two_bytes() {
-        let grams = collect_sparse_grams(b"ab");
-        assert_eq!(grams.len(), 1);
-        assert_eq!(grams[0], NGram::from_bytes(b"ab"));
-    }
-
-    #[test]
-    fn test_three_bytes() {
-        let grams = collect_sparse_grams(b"abc");
-        assert!(grams.len() >= 2);
-        assert_eq!(grams[0], NGram::from_bytes(b"ab"));
-        assert_eq!(grams[1], NGram::from_bytes(b"bc"));
-    }
-
-    #[test]
-    fn test_gram_lengths_bounded() {
-        let input = b"self.reset_states(the_quick_brown_fox_jumps";
-        let grams = collect_sparse_grams(input);
-        for gram in &grams {
-            assert!(gram.len() >= 2, "gram too short: {gram:?}");
-            assert!(gram.len() <= MAX_SPARSE_GRAM_SIZE as usize, "gram too long: {gram:?}");
-        }
-    }
-
-    #[test]
-    fn test_produces_longer_grams() {
-        let grams = collect_sparse_grams(b"self.reset_states(");
-        assert!(grams.iter().any(|g| g.len() > 2));
-    }
-
-    #[test]
-    fn test_max_gram_size_boundary() {
-        let grams = collect_sparse_grams(b"abcdefgh");
-        for gram in &grams {
-            assert!(gram.len() <= MAX_SPARSE_GRAM_SIZE as usize);
-        }
-    }
-
-    #[test]
-    fn test_repeated_bytes() {
-        let grams = collect_sparse_grams(b"aaaaaaaaaa");
-        assert!(grams.iter().filter(|g| g.len() == 2).count() >= 9);
-    }
-
-    #[test]
-    fn test_gram_count_scales_linearly() {
-        let input: Vec<u8> = (0..1000).map(|i| (i % 256) as u8).collect();
-        let grams = collect_sparse_grams(&input);
-        assert!(grams.len() >= input.len() - 1);
-        assert!(grams.len() <= input.len() * 3);
-    }
-
-    // -- Equivalence: scan vs deque --
-
-    #[test]
-    fn test_scan_equivalence_small() {
-        for input in [b"" as &[u8], b"x", b"ab", b"abc", b"abcdefgh", b"abcdefghi"] {
-            assert_eq!(
-                collect_to_vec(input, collect_sparse_grams_deque),
-                collect_to_vec(input, collect_sparse_grams_scan),
-                "mismatch on {:?}",
-                std::str::from_utf8(input).unwrap_or("?")
-            );
-        }
-    }
-
-    #[test]
-    fn test_scan_equivalence_hello_world() {
-        let input = b"hello world";
-        assert_eq!(
-            collect_to_vec(input, collect_sparse_grams_deque),
-            collect_to_vec(input, collect_sparse_grams_scan),
-        );
-    }
-
-    #[test]
-    fn test_scan_equivalence_large() {
-        let input: Vec<u8> = (0..1000).map(|i| (i % 256) as u8).collect();
-        assert_eq!(
-            collect_to_vec(&input, collect_sparse_grams_deque),
-            collect_to_vec(&input, collect_sparse_grams_scan),
-        );
-    }
-
-    #[test]
-    fn test_scan_equivalence_source_code() {
-        let input = include_bytes!("lib.rs");
-        assert_eq!(
-            collect_to_vec(input, collect_sparse_grams_deque),
-            collect_to_vec(input, collect_sparse_grams_scan),
-        );
-    }
-
-}
-
